07/07

+++ 09_user_add.c
@@ -1,12 +1,12 @@
 #include<stdio.h>
 void main()
 {
-	int x,y,z;
+	float x,y,z;
 	printf("this is addiotion app \n");
 	printf("enter value of x = ");
-	scanf("%d",&x);
+	scanf("%f",&x);
 	printf("enter value of y = ");
-	scanf("%d",&y);
+	scanf("%f",&y);
 	z = x+y;	
-	printf("addition = %d",z);
+	printf("addition = %.1f",z);
 }
